moved impl CustomMessage for Vec
