+++ p2p/src/lib.rs
@@ -13,3 +13,7 @@ mod priority;
 pub use self::node::{Direction, Node, NodeConfig, NodeHandle, NodeNotification, PeerInfo};
 pub use self::peer::{CustomMessage, PeerID, PeerLink, PeerMessage, PeerNotification};
 pub use self::priority::Priority;
+
+pub mod reexport {
+    pub use bytes::{BytesMut, Bytes, BufMut, Buf};
+}
